# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "wecom KB API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/v1/admin/roles": {
      "get": {
        "tags": [
          "admin"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "name": {
                            "type": "string",
                            "minLength": 1
                          },
                          "description": {
                            "type": "string",
                            "default": ""
                          },
                          "system": {
                            "type": "boolean"
                          },
                          "permissions": {
                            "type": "array",
                            "items": {
                              "type": "string",
                              "enum": [
                                "docs.read",
                                "docs.create",
                                "docs.edit",
                                "docs.publish",
                                "docs.delete",
                                "docs.restore",
                                "blocks.edit",
                                "fields.edit",
                                "scripts.edit",
                                "notes.write",
                                "notes.moderate",
                                "suggestions.review",
                                "suggestions.apply",
                                "sources.manage",
                                "connectors.manage",
                                "users.manage",
                                "roles.manage",
                                "audit.read",
                                "system.admin"
                              ]
                            }
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "system",
                          "permissions"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "items"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 40
                  },
                  "description": {
                    "type": "string",
                    "default": ""
                  },
                  "permissions": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "docs.read",
                        "docs.create",
                        "docs.edit",
                        "docs.publish",
                        "docs.delete",
                        "docs.restore",
                        "blocks.edit",
                        "fields.edit",
                        "scripts.edit",
                        "notes.write",
                        "notes.moderate",
                        "suggestions.review",
                        "suggestions.apply",
                        "sources.manage",
                        "connectors.manage",
                        "users.manage",
                        "roles.manage",
                        "audit.read",
                        "system.admin"
                      ]
                    }
                  }
                },
                "required": [
                  "name",
                  "permissions"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string",
                      "minLength": 1
                    },
                    "description": {
                      "type": "string",
                      "default": ""
                    },
                    "system": {
                      "type": "boolean"
                    },
                    "permissions": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "docs.read",
                          "docs.create",
                          "docs.edit",
                          "docs.publish",
                          "docs.delete",
                          "docs.restore",
                          "blocks.edit",
                          "fields.edit",
                          "scripts.edit",
                          "notes.write",
                          "notes.moderate",
                          "suggestions.review",
                          "suggestions.apply",
                          "sources.manage",
                          "connectors.manage",
                          "users.manage",
                          "roles.manage",
                          "audit.read",
                          "system.admin"
                        ]
                      }
                    },
                    "auditId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "system",
                    "permissions",
                    "auditId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/roles/{id}": {
      "patch": {
        "tags": [
          "admin"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 40
                  },
                  "description": {
                    "type": "string",
                    "default": ""
                  },
                  "permissions": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "docs.read",
                        "docs.create",
                        "docs.edit",
                        "docs.publish",
                        "docs.delete",
                        "docs.restore",
                        "blocks.edit",
                        "fields.edit",
                        "scripts.edit",
                        "notes.write",
                        "notes.moderate",
                        "suggestions.review",
                        "suggestions.apply",
                        "sources.manage",
                        "connectors.manage",
                        "users.manage",
                        "roles.manage",
                        "audit.read",
                        "system.admin"
                      ]
                    }
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string",
                      "minLength": 1
                    },
                    "description": {
                      "type": "string",
                      "default": ""
                    },
                    "system": {
                      "type": "boolean"
                    },
                    "permissions": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "docs.read",
                          "docs.create",
                          "docs.edit",
                          "docs.publish",
                          "docs.delete",
                          "docs.restore",
                          "blocks.edit",
                          "fields.edit",
                          "scripts.edit",
                          "notes.write",
                          "notes.moderate",
                          "suggestions.review",
                          "suggestions.apply",
                          "sources.manage",
                          "connectors.manage",
                          "users.manage",
                          "roles.manage",
                          "audit.read",
                          "system.admin"
                        ]
                      }
                    },
                    "auditId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "system",
                    "permissions",
                    "auditId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "auditId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "ok",
                    "auditId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}